+++ da_vinci-cdk/da_vinci_cdk/framework_stacks/services/event_bus/runtime/bus.py
@@ -1,4 +1,4 @@
-'''Event Bus runtime'''
+'''Event Bus Router'''
 import json
 import logging