+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Add support for converting nested object bodies when parent object body is converted to a dict
 - Fix event bus object mismatched type exception message
 - Simplify event bus object schema class
+- Fix event bus object body value reference bug
 
 ### 2024.12.2 (Latest)
 - Add update support to DynamoDB Item CDK construct (Infra